Home CoreData logics moved to its related protocol
