+++ BluSwiftUI/App/TabGroup/HomeGroup/Home/ViewModel/HomeViewModel.swift
@@ -10,17 +10,21 @@ import CoreData
 
 @MainActor
 final class HomeViewModel: ObservableObject {
+    // MARK: - Properties
     @Published private(set) var transfers: HomeResponseModel = []
     @Published private(set) var isGetMoreItemsLoading: Bool = false
     
     private var page: Int = 1
     
     let homeAPIProtocol: HomeAPIProtocol!
+    let homeCoreDataProtocol: HomeCoreDataProtocol!
     
-    init(homeAPIProtocol: HomeAPIProtocol = HomeAPI()) {
+    init(homeAPIProtocol: HomeAPIProtocol = HomeAPI(), homeCoreDataProtocol: HomeCoreDataProtocol = HomeCoreData()) {
         self.homeAPIProtocol = homeAPIProtocol
+        self.homeCoreDataProtocol = homeCoreDataProtocol
     }
-    
+    // MARK: - Methods
+    // MARK: APIs
     func getTransferList() async {
         do {
             if let data = try await homeAPIProtocol.getTransferList(page: page) {
@@ -58,34 +62,12 @@ final class HomeViewModel: ObservableObject {
     func checkIfLastItem(_ currentHomeModel: HomeModel) -> Bool {
         return currentHomeModel.id == transfers.last?.id
     }
-    
+    // MARK: Core Data
     func checkIsInFavorites(_ favorites: FetchedResults<TransferCDModel>, homeModel: HomeModel) -> Bool {
-        favorites.contains(where: {$0.fullName == homeModel.person.fullName})
-    }
-    
-    private func addToFavorites(context: NSManagedObjectContext, homeModel: HomeModel) {
-        let newFavoriteItem = TransferCDModel(context: context)
-        newFavoriteItem.id = homeModel.id
-        newFavoriteItem.fullName = homeModel.person.fullName
-        newFavoriteItem.email = homeModel.person.email
-        newFavoriteItem.url = homeModel.person.avatar
-        
-        try? context.save()
-    }
-    
-    private func removeFromFavorites(_ favorites: FetchedResults<TransferCDModel>, context: NSManagedObjectContext, homeModel: HomeModel) {
-        if let item = favorites.first(where: {$0.fullName == homeModel.person.fullName}) {
-            context.delete(item)
-            
-            try? context.save()
-        }
+        homeCoreDataProtocol.checkIsInFavorites(favorites, homeModel: homeModel)
     }
     
     func favoriteButtonTapped(favorites: FetchedResults<TransferCDModel>, context: NSManagedObjectContext, homeModel: HomeModel) {
-        if checkIsInFavorites(favorites, homeModel: homeModel) {
-            removeFromFavorites(favorites, context: context, homeModel: homeModel)
-        } else {
-            addToFavorites(context: context, homeModel: homeModel)
-        }
+        homeCoreDataProtocol.favoriteButtonTapped(favorites: favorites, context: context, homeModel: homeModel)
     }
 }
